+++ sdk/python/__init__.py
@@ -1,15 +1,19 @@
 from .agent_bus_sdk import (
     AgentBusClient,
     AgentBusError,
+    ROOM_EVENT_TYPES,
     agent_model,
     replay_room_events,
     room_event_bundle,
+    validate_room_event_bundle,
 )
 
 __all__ = [
     "AgentBusClient",
     "AgentBusError",
+    "ROOM_EVENT_TYPES",
     "agent_model",
     "replay_room_events",
     "room_event_bundle",
+    "validate_room_event_bundle",
 ]